+++ database/upgrades/00-latest.sql
@@ -1,4 +1,4 @@
--- v0 -> v3 (compatible with v1+): Latest schema
+-- v0 -> v4 (compatible with v1+): Latest schema
 CREATE TABLE bot (
     username     TEXT PRIMARY KEY NOT NULL,
     displayname  TEXT NOT NULL,
@@ -37,3 +37,5 @@ CREATE TABLE policy_server_signature (
 
     PRIMARY KEY (event_id)
 );
+
+CREATE TABLE otks_need_reset(user_id TEXT PRIMARY KEY);