+++ policyeval/policyserver_check.go
@@ -1,4 +1,4 @@
-//go:build goexperiment.jsonv2
+//go:build goexperiment.jsonv2 || go1.27
 
 package policyeval
 
@@ -19,8 +19,9 @@ import (
 )
 
 type BlockRecommendation struct {
-	Error error            `json:"error,omitempty"`
-	Match policylist.Match `json:"match,omitempty"`
+	Error        error            `json:"error,omitempty"`
+	Match        policylist.Match `json:"match,omitempty"`
+	DisplayError string           `json:"display_error,omitempty"`
 }
 
 func (ps *PolicyServer) getRecommendation(
@@ -29,27 +30,38 @@ func (ps *PolicyServer) getRecommendation(
 	roomVersion id.RoomVersion,
 	evaluator *PolicyEvaluator,
 	isOrigin, isLegacyCheck bool,
-) (string, *BlockRecommendation) {
+) (string, *BlockRecommendation, error) {
 	if pdu.Sender == evaluator.Bot.UserID || evaluator.Admins.Has(pdu.Sender) {
-		return "admin", nil
+		return "admin", nil, nil
+	}
+	err := evaluator.loadingComplete.Wait(ctx)
+	if err != nil {
+		return "", nil, err
 	}
 	watchedLists := evaluator.GetWatchedLists()
 	match := evaluator.Store.MatchUser(watchedLists, pdu.Sender)
 	if match != nil {
-		rec := match.Recommendations().BanOrUnban
-		if rec != nil && rec.Recommendation != event.PolicyRecommendationUnban {
-			return "", &BlockRecommendation{Match: match}
+		recs := match.Recommendations()
+		if recs.BanOrUnban != nil && recs.BanOrUnban.Recommendation != event.PolicyRecommendationUnban {
+			return "", &BlockRecommendation{Match: match, DisplayError: "You are banned in this room"}, nil
+		} else if recs.Mute != nil {
+			return "", &BlockRecommendation{Match: match, DisplayError: fmt.Sprintf("You are muted in this room: %s", recs.Mute.Reason)}, nil
 		}
 	}
 	match = evaluator.Store.MatchServer(watchedLists, pdu.Sender.Homeserver())
 	if match != nil {
-		rec := match.Recommendations().BanOrUnban
-		if rec != nil && rec.Recommendation != event.PolicyRecommendationUnban {
-			return "", &BlockRecommendation{Match: match}
+		recs := match.Recommendations()
+		if recs.BanOrUnban != nil && recs.BanOrUnban.Recommendation != event.PolicyRecommendationUnban {
+			return "", &BlockRecommendation{Match: match, DisplayError: "You are banned in this room"}, nil
+		} else if recs.Mute != nil {
+			return "", &BlockRecommendation{Match: match, DisplayError: fmt.Sprintf("You are muted in this room: %s", recs.Mute.Reason)}, nil
 		}
 	}
 	if pdu.StateKey == nil && !pdu.VerifyContentHash() {
-		return "", &BlockRecommendation{Error: fmt.Errorf("mismatching content hash")}
+		return "", &BlockRecommendation{
+			Error:        fmt.Errorf("mismatching content hash"),
+			DisplayError: "Event is malformed",
+		}, nil
 	}
 	if evaluator.protections != nil {
 		clientEvt, err := pdu.ToClientEvent(roomVersion)
@@ -57,7 +69,10 @@ func (ps *PolicyServer) getRecommendation(
 			zerolog.Ctx(ctx).Err(err).
 				Stringer("room_id", pdu.RoomID).
 				Msg("Failed to convert PDU to client event")
-			return "", &BlockRecommendation{Error: fmt.Errorf("failed to convert PDU to client event: %w", err)}
+			return "", &BlockRecommendation{
+				Error:        fmt.Errorf("failed to convert PDU to client event: %w", err),
+				DisplayError: "Event is malformed",
+			}, nil
 		}
 		if parseErr := clientEvt.Content.ParseRaw(clientEvt.Type); parseErr != nil {
 			evaluator.Bot.Log.Err(parseErr).
@@ -89,12 +104,15 @@ func (ps *PolicyServer) getRecommendation(
 				}
 				zerolog.Ctx(ctx).Trace().Bool("spam", rec).Str("protection", name).Msg("Evaluated protection")
 				if rec {
-					return "", &BlockRecommendation{Error: fmt.Errorf("protections rejected event")}
+					return "", &BlockRecommendation{
+						Error:        fmt.Errorf("protections rejected event"),
+						DisplayError: "This message has been rejected as probable spam",
+					}, nil
 				}
 			}
 		}
 	}
-	return "no reason to disallow", nil
+	return "no reason to disallow", nil, nil
 }
 
 const PolicyServerKeyID id.KeyID = "ed25519:policy_server"
@@ -106,13 +124,13 @@ func (ps *PolicyServer) HandleSign(
 	evt *pdu.PDU,
 	evaluator *PolicyEvaluator,
 	originServer string,
-) error {
+) (string, error) {
 	if ps.SigningKey == nil {
-		return errors.New("policy server is not configured with a signing key")
+		return "", errors.New("policy server is not configured with a signing key")
 	}
 	evtID, err := evt.GetEventID(roomVersion)
 	if err != nil {
-		return fmt.Errorf("failed to calculate event ID: %w", err)
+		return "", fmt.Errorf("failed to calculate event ID: %w", err)
 	}
 	log := zerolog.Ctx(ctx).With().
 		Stringer("room_id", evt.RoomID).
@@ -120,18 +138,18 @@ func (ps *PolicyServer) HandleSign(
 		Logger()
 	sig, err := ps.DB.PSSignature.Get(ctx, evtID)
 	if err != nil {
-		return fmt.Errorf("failed to fetch signature from database: %w", err)
+		return "", fmt.Errorf("failed to fetch signature from database: %w", err)
 	} else if sig != nil {
 		log.Trace().
 			Time("cached_at", sig.CreatedAt.Time).
 			Str("signature", sig.Signature).
 			Msg("Using cached result for sign request")
 		evt.AddSignature(ps.Federation.ServerName, PolicyServerKeyID, sig.Signature)
-		return nil
+		return "", nil
 	}
 
 	log.Trace().Any("event", evt).Msg("Checking event received by policy server")
-	allowReason, blockRec := ps.getRecommendation(
+	allowReason, blockRec, err := ps.getRecommendation(
 		ctx,
 		evt,
 		roomVersion,
@@ -139,20 +157,23 @@ func (ps *PolicyServer) HandleSign(
 		originServer == evt.Sender.Homeserver(),
 		originServer == fakeLegacyCheckServerName,
 	)
+	if err != nil {
+		return "", err
+	}
 	if blockRec != nil {
 		// Don't sign spam events
 		log.Debug().Any("recommendation", blockRec).Msg("Event rejected for spam")
-		return nil
+		return blockRec.DisplayError, nil
 	}
 	log.Trace().Str("allow_reason", allowReason).Msg("Event accepted")
 
 	err = evt.Sign(roomVersion, ps.Federation.ServerName, PolicyServerKeyID, ps.SigningKey.Priv)
 	if err != nil {
-		return fmt.Errorf("failed to add signature to PDU: %w", err)
+		return "", fmt.Errorf("failed to add signature to PDU: %w", err)
 	}
 	newSig, ok := evt.Signatures[ps.Federation.ServerName][PolicyServerKeyID]
 	if !ok {
-		return errors.New("failed to retrieve signature after signing")
+		return "", errors.New("failed to retrieve signature after signing")
 	}
 	err = ps.DB.PSSignature.Put(ctx, &database.PSSignature{
 		EventID:   evtID,
@@ -160,9 +181,9 @@ func (ps *PolicyServer) HandleSign(
 		CreatedAt: jsontime.UnixMilliNow(),
 	})
 	if err != nil {
-		return fmt.Errorf("failed to store signature in database: %w", err)
+		return "", fmt.Errorf("failed to store signature in database: %w", err)
 	}
-	return nil
+	return "", nil
 }
 
 func (ps *PolicyServer) getSigningKey(serverName string, keyID id.KeyID, minValidUntil time.Time) (id.SigningKey, time.Time, error) {
@@ -187,7 +208,7 @@ func (ps *PolicyServer) HandleLegacyCheck(
 		log.Trace().Msg("Valid signature from self, short-circuiting legacy check")
 		return true, nil
 	}
-	err = ps.HandleSign(ctx, roomVersion, pdu, evaluator, fakeLegacyCheckServerName)
+	_, err = ps.HandleSign(ctx, roomVersion, pdu, evaluator, fakeLegacyCheckServerName)
 	if err != nil {
 		return false, err
 	}