+++ policylist/policy.go
@@ -29,11 +29,15 @@ type Match []*Policy
 
 type Recommendations struct {
 	BanOrUnban *Policy
+	Mute       *Policy
 }
 
 func (r Recommendations) String() string {
 	if r.BanOrUnban != nil {
 		return string(r.BanOrUnban.Recommendation)
+	}
+	if r.Mute != nil {
+		return string(r.Mute.Recommendation)
 	}
 	return ""
 }
@@ -42,9 +46,21 @@ func (r Recommendations) String() string {
 func (m Match) Recommendations() (output Recommendations) {
 	for _, policy := range m {
 		switch policy.Recommendation {
-		case event.PolicyRecommendationBan, event.PolicyRecommendationUnban, event.PolicyRecommendationUnstableTakedown:
+		case event.PolicyRecommendationUnban:
+			if output.BanOrUnban == nil || output.BanOrUnban.RoomID == policy.RoomID {
+				output.BanOrUnban = policy
+			}
+			if output.Mute != nil && output.Mute.RoomID == policy.RoomID {
+				output.Mute = nil
+			}
+			return
+		case event.PolicyRecommendationBan, event.PolicyRecommendationUnstableTakedown:
 			if output.BanOrUnban == nil {
 				output.BanOrUnban = policy
+			}
+		case event.PolicyRecommendationMute:
+			if output.Mute == nil && (output.BanOrUnban == nil || output.BanOrUnban.Recommendation != event.PolicyRecommendationUnban) {
+				output.Mute = policy
 			}
 		}
 	}